Module rust_ipv8::networking

This module handles all the network IO and allows senders to send and notifies receivers of recieved messages.

Modules

address
Module containing structs for working with network addresses

Structs

NetworkReceiver
A NetworkReceiver is another (see NetworkSender) wrapper for a socket.

NetworkSender
A NetworkSender is a wrapper for a sending udp socket.

Traits

Receiver
Any struct implementing this method can become a receiver of incoming network packets. under normal operation, only the IPV8 struct should be a receiver of these and it should distribute it through its CommunityRegistry to communities