Module rust_ipv8::payloads

This module has all of the original py-ipv8 payloads in it to be put into packets and sent in a way such that py-ipv8 can interpret what we are sending.

Modules

binmemberauthenticationpayload
The BinMemberAuthenticationPayload module contains the structs and methods necessary for representing serializing and deserializing BinMemberAuthenticationPayloads

connectiontype
Represents the type of connection two communities have.

introductionrequestpayload
Module containg the IntroductionRequestPayload relevant code

introductionresponsepayload
Sent by the base Community to start a direct connection between two Peers

puncturepayload
Payload sent by the receiver of a PunctureRequestPayload to actually puncture a hole (NAT puncturing)

puncturerequestpayload
Payload sent before performing NAT puncturing

timedistributionpayload
Send as a part of most Packets to drive the Lamport clock

Traits

Ipv8Payload
Used to recognize payloads. Does not have any members yet, (though already useful as a marker trait).